Moved registration to inside sensor_processing node
