--- files/registration.csv
@@ -1,7 +1,0 @@
--1.059989395141601562e+02
--9.256348419189453125e+01
-1.304706726074218750e+02
-1.000000000000000000e+00
-0.000000000000000000e+00
-0.000000000000000000e+00
-0.000000000000000000e+00
